Rendicion en proceso. Actualizacion de AbmFactura testeado

+++ scriptsMsSqlServer/Stored Procedures/WEST_WORLD.ItemDelete.sql
@@ -3,9 +3,8 @@ CREATE PROCEDURE WEST_WORLD.ItemDelete
 @IDITEM bigint
 
 AS
-	
-	DELETE FROM WEST_WORLD.Item
-	WHERE numeroFactura = @NUMEROFACTURA and idItem = @IDITEM
-
-RETURN 0
-
+	IF((SELECT COUNT(idItem) FROM WEST_WORLD.Item WHERE numeroFactura = @NUMEROFACTURA) > 1)
+		DELETE FROM WEST_WORLD.Item
+		WHERE numeroFactura = @NUMEROFACTURA and idItem = @IDITEM
+	ELSE
+		RAISERROR('La factura no puede quedar sin items. Puede actualizar el seleccionado', 15, 2)
